Create AdPage.java
An AdPage will be poped before actualy taking the quiz.  
There are 2 parts on this page, Part A is Advertisement Part, User can click the link to see more about the advertisement, for here, we only use www.pitt.edu. 
Part B is StartQuiz button, this button can only be pressed after 5 seconds for user which is used to force the user watch the ad and let them prepare. After 5 seconds, button can be pressed to take the quiz.